DB_Tables_Raw/missing_requirements.txt
autocommand==2.2.2
backports.tarfile==1.2.0
bleach==6.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
kaggle==1.7.4.5
kiwisolver==1.4.9
matplotlib==3.10.8
maturin==1.9.6
more-itertools==10.3.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pip==25.3
platformdirs==4.2.2
protobuf==6.33.0
psycopg2-binary==2.9.11
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.67.1
typeguard==4.3.0
typing-extensions==4.12.2
tzdata==2025.3
urllib3==2.5.0
webencodings==0.5.1
wheel==0.45.1
zipp==3.19.2
